🎨 style(server.ts): change port variable case from lowercase port to uppercase PORT ✨ feat(server.ts): add support for process.env.PORT environment variable The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🐛 fix(quiz.json): add missing logoUrl and schoolMeta properties
The logoUrl and schoolMeta properties were missing from the quiz.json file. These properties are necessary for displaying the school logos and metadata on the front-end.

🐛 fix(next.config.js): update remotePatterns hostname
The remotePatterns hostname in next.config.js was pointing to the wrong URL. This has been updated to the correct URL.

🚀 chore(package.json): add react-multi-carousel dependency
The react-multi-carousel package has been added to the project dependencies to enable

🎨 style(practitioner.jsx, scientist.jsx): replace UniversityMatch with SchoolCarousel
The UniversityMatch component has been replaced with the more versatile SchoolCarousel component. This change allows for the display of multiple matched schools instead of just one. The matchedSchool variable has been renamed to matchedSchools to reflect the change.