🔥 remove(Stats.js, Tabs.js, UniversityMatch.js): replace p tag with div tag The p tag was replaced with a div tag to avoid rendering issues with nested HTML tags.
🚀 feat(getQuizJSON.js): add function to read quiz data from file
The getQuizJSON function reads the quiz data from a JSON file and returns the results array.

✨ feat(analyst.jsx): add getStaticProps to fetch quiz data
The getStaticProps function fetches the quiz data using the getQuizJSON function and returns the results and schools data as props to the AnalystPage component. The CappexFormSection component was commented out as it is not currently being used.